+++ packages/cat-sdk/src/features/cat20/mint/closedMinter.ts
@@ -98,7 +98,7 @@ export const mintClosedMinterToken = createFeatureWithDryRun(async function(
   )
   const utxos = await provider.getUtxos(address)
 
-  closedMinter.bindToUtxo(minterUtxo)
+  closedMinter.bindToUtxo({...minterUtxo, txHashPreimage: toHex(new Transaction(spentMinterTxHex).toTxHashPreimage())})
   closedMinter.state = minterState
 
   const nextMinter = closedMinter.next(closedMinter.state)