+++ packages/integration-examples/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @opcat-labs/integration-examples
 
+## 2.1.0
+
+### Minor Changes
+
+- back to genesis now trace to genesis contract
+- add dryRun for cat-sdk features
+- add mergeSendToken feature for cat20
+- lots of bugfixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @opcat-labs/cat-sdk@2.1.0
+  - @opcat-labs/scrypt-ts-opcat@2.1.0
+
 ## 2.0.2
 
 ### Patch Changes